fix: [io/fileinfoquery]Paste the file into the USB drive, but the file is not displayed
When creating a new file, insert it into the cache and immediately create a new itemfiledata. When creating fileinfo in itemfiledata, you may not have obtained the cached fileinfo yet. Therefore, you have created a new one and will asynchronously query the new one. Will query the previous one instead of all subsequent ones. All fileinfo attributes are invalid. Each fileinfo can be modified to query, and a corresponding identifier can be added to the successfully queried signal. The slot function of the received signal interprets whether it is the same identifier for processing.

Log: Paste the file into the USB drive, but the file is not displayed
Bug: https://pms.uniontech.com/bug-view-200183.html