fix: [menu/asyncfileinfout]The right mouse button of the file manager is stuck, and the ut modification of asyncfileinfo
Fileinfo was created in the updateState function in the FileOperatorMenuScene class. Modify here to use the selected fileinfo. If the selected file passed in does not match the selected fieldinfo, add the fileinfo of the last selected file. UT modification is processed using asynchronous threads.

Log: The right mouse button of the file manager is stuck, and the ut modification of asyncfileinfo