fix: filemanager window is showed in Initializing.
It is incorrect that showing a window that is Initializing.

When using the D-Bus interface to create a file dialog, the internal process of the file manager calls the show method, causing the file dialog to be shown and stealing the window focus upon creation. However, after the file dialog is created, it is then immediately hidden without being showed.

This results in only the creation of the file dialog, while causing other windows to lose focus.

Add property _dfm_Disable_RestoreWindowState_ to disable restore window state.

Log: 

Bug: https://pms.uniontech.com/bug-view-22564.html
Bug: https://pms.uniontech.com/bug-view-204221.html
 
Influence: 文管窗口状态，文件选择对话框的呼出状态