fix: [io]Failure to pop up in a timely manner indicates that files exceeding 4G cannot be copied
There is a check in front of the copy logic of v5 that the vfat format has been manipulated for 4g and has not been moved to v6. Check for adding vfat file operation 4g.

Log: Failure to pop up in a timely manner indicates that files exceeding 4G cannot be copied
Bug: https://pms.uniontech.com/bug-view-204799.html