fix: [filesortwork]File Manager Crash
The asynchronous thread checknamefilters called the fileinfo interface of fileitemdata, causing both the main thread and asynchronous thread to execute the construction of fileinfo simultaneously, resulting in one of the fileinfo being destructed. Modify the use of other methods to determine whether the file is a directory when using the checknamefilters function on asynchronous threads.

Log: File Manager Crash