fix: [io]Unable to iterate over files
When the same url is opened using multiple windows at the same time, there is only one rootinfo corresponding to multiple models, but one model corresponds to one Iterator. All iteration signals of the Iterator are connected to rootinfo, and the correct model is not found. Add a token to find the correct model.There is only one iteration thread processing in the same directory, while other window iterations use the data from the current iteration.

Log: Unable to iterate over files
Bug: https://pms.uniontech.com/bug-view-205505.html