fix: [titlebar]The address bar folder title name is incorrect
In the trash file, the gio trash mechanism is used, and files with the same name will be added with. index. suffix after the file name. So the address in the titl column shows an error. If the schema is trash, use the displayname in fileinfo.

Log: The address bar folder title name is incorrect
Bug: https://pms.uniontech.com/bug-view-205729.html