fix: [mount] device is mounted at /media/root
device is mounted at /media/root/$DISKNAME while root user is not
loggined. the auto mount operation is dealed in server process, and when
an admin-dfm is launched before, the server which service for admin-dfm
is launched too. and device can be mounted by the admin-server when
device plug in.

Log: fix issue about auto mount.

Bug: https://pms.uniontech.com/bug-view-205301.html