fix: [fileinfo]Thumbnail keeps refreshing
Asynchronous threads are used every time in fileicon in fileinfo to obtain thumbnails. The desktop and civil servants call the refresh interface of fileinfo upon receiving the completion of thumbnail generation. This is caused by the thread to continue looping while trying to obtain the generated thumbnail. Modify the fileicon of fileinfo by first reading the thumbnail and returning it if there is one.

Log: Thumbnail keeps refreshing
Bug: https://pms.uniontech.com/bug-view-206629.html