fix: [trashfileinfo]Application files (. desktop files) cannot be restored properly after deletion
There was an error in obtaining the original name of the deleted file while checking the file name. If it was a desktop file in trashfileinfo, it returned the displayed name of the file, so the error occurred. Modifying the file name should return the file source file name.

Log: Application files (.desktop files) cannot be restored properly after deletion
Bug: https://pms.uniontech.com/bug-view-206719.html