fix: [recent/fileinfo]File manager opens, leaves, closes, and occasionally crashes after recent use
When closing, the system is destructing. Here, the Static variable in the fileinfo of the file information has been modified, and the vector access again is out of bounds. Modify each attribute of fileinfo and pass it in yourself. Add the stop flag in the identifier worker and exit once stopped.

Log: File manager opens, leaves, closes, and occasionally crashes after recent use