fix: [workspace]Quickly switch directories, file manager crashes
Here, the sorting thread called fileitemdata's fileinfo to obtain the mimetype of the file, and the main thread may also have called fileinfo creation at the same time, causing fileinfo to be destructed and causing a crash. Modify the fileinfo in fileitemdata to obtain and not create it. The fileinfo of the specified item in the data is created on the main thread.

Log: Quickly switch directories, file manager crashes