fix: [io/thumbnail]Thumbnail display error
The source file of the linked file is a remote file, displaying a thumbnail error. When obtaining the file's properties in the thumbnail helper class, the default was returned (the asynchronous file information query at this time has not been completed). When the asynchronous file information acquisition is completed, refresh the thumbnail flag in fileinfo.

Log: Thumbnail display error