fix: [io/taskwidget]Error in copying Progress bar information
Because when copying to a USB drive, the file information is asynchronous and the attributes of the file have not been obtained yet, the displayed data is not accurate. Modify the timing to obtain file attributes, and if it is correct, it will not be obtained.

Log: Error in copying Progress bar information
Bug: https://pms.uniontech.com/bug-view-209653.html