chore: [io/iterator]dde-file-manager adapts the initEnumerator interface added by dfmio.
dde-file-manager needs to know that the iterator initialization is complete when iterating over files. Modifications: 1. Add initIterator interface to the AbstractDirIterator class. 2. Add initIterator interface implementations to the LocalDirIterator, VaultFileIterator and SmbShareIterator classes, calling the dfmio class. 3. DEnumerator's initEnumerator interface.3. Modify the TraversalDirThreadManager class to call initialization in the thread execution, and send the iteratorInitFinished signal when it is done

Log: dde-file-manager adapts the initEnumerator interface added by dfmio.