fix: [io/proxyfileinfo]The masteredmedia keeps flashing
The file information for the files in the masteredmedia is an inherited proxy file information class. When you create the file information for the file in the masteredmedia, you create the proxy file information, which in turn is asynchronous file information, and then you set up the proxy and call the proxy's refresh interface. In the thumbnail creation task, after creating the file information of the file in the masteredmedia, when leaving to get its properties, it is invalid file information, resulting in the thumbnail of the file being created all the time. Modification: 1. When setting up the proxy file information, do not call the proxy file information's refresh. 2. Add an interface to remove the notification url in the asynchronous file information class. 3. Remove the notification url in the proxy when the proxy file information class destructs.

Log: The masteredmedia keeps flashing
Bug: https://pms.uniontech.com/bug-view-210839.html