fix: [vault]No semi-transparent drawing after cut files in vault.
Optimize ctrl+c to get rid of the previous fileinfo creation to get the inode and record it in the list of cut files. When judging the semi-transparent drawing, I just use url to judge, and the virtual directory fails to judge. Modify the file that can be redirected, use the absolute path of its local file to construct the url, and judge the cut url.

Log: No semi-transparent drawing after cut files in vault.
Bug: https://pms.uniontech.com/bug-view-211525.html