fix: [io]Multi-tiered mass file copying to vault within SMB is very slow
Here the copy is not slow, it's a problem when counting write speeds before there is a tid for the thread, read the io corresponding to this tid in the system and get the write size error. Modify to add jogflag to set up for their own statistics copy progress.

Log: Multi-tiered mass file copying to vault within SMB is very slow
Bug: https://pms.uniontech.com/bug-view-212635.html