fix: [trash]Incorrect dde-desktop trash icon
When initializing TrashCoreEventSender, the FileUtils::trashIsEmpty interface is called to set the flag isEmpty in TrashCoreEventSender. at this time, trashinfo is not yet registered, so the TrashCoreEventSender flag isEmpty is true, clearing the trash will not send the trash is empty event. Modify to register trashinfo first then initialize TrashCoreEventSender.

Log: Incorrect dde-desktop trash icon
Bug: https://pms.uniontech.com/bug-view-213989.html