fix: [io]Peripheral paste the same name directory selection coexistence, file sorting is not correct
Asynchronous file information is created when a file is received from the monitor to create a file, multiple queries are executed, and caching of file attributes is performed in multiple threads, but the corresponding dfmfileinfo is different. So accessing file attributes at this point is incorrect.

Log: Peripheral paste the same name directory selection coexistence, file sorting is not correct need to refresh manually
Bug: https://pms.uniontech.com/bug-view-217143.html