fix: [fileinfo]Incorrect icon for dde-desktop's file
When the desktop is a remote environment, when the file manager creates a new word document, it performs two actions, a touchfile, in writing the template information. In the absence of writing template information, the desktop's asynchronous file information query back icon is a text document, received a signal to change the file information in the go to refresh fileinfo, because at this time is caching the file attributes, so the return is not executed. Modification in the cache information is possible to perform asynchronous file query.

Log: Incorrect icon for dde-desktop's file
Bug: https://pms.uniontech.com/bug-view-217907.html