fix: [io]After sending a folder to the desktop with smb service mounted, the link target file cannot be found when previewing it with space.
After reboot, it is not mounted, and when it is mounted again, the desktop does not listen to the protocol device, so it doesn't update the fileinfo. so there is something wrong with the attribute.

Log: After sending a folder to the desktop with smb service mounted, the link target file cannot be found when previewing it with space.
Bug: https://pms.uniontech.com/bug-view-219383.html