fix: [workspace]Creating a wps file in red disk space doesn't show up, it needs to be refreshed before it shows up
In the current interface, safely uninstalled the USB flash drive, did not remove the monitor, used the previous one, and has been using the monitor in the cache since then. This resulted in the files not being monitored.

Log: Creating a wps file in red disk space doesn't show up, it needs to be refreshed before it shows up
Bug: https://pms.uniontech.com/bug-view-221223.html