fix: [io]The red disk monitor is not working.
When the file manager went to the red disk, unmounted and mounted the red disk again, the monitoring did not stop, causing the underlying gio to monitor the same directory as the last time. The monitor is actively stopped when exiting the current directory, and also when removing the monitor.

Log: The red disk monitor is not working.