fix: [io]Optical drive list view, after mounting, it takes a while for the file names inside the CD-ROM to be displayed in their entirety
Asynchronous file information is not yet complete, go get kItemFileBaseNameRole file information. This is because the cache doesn't use the parent class fileinfo to get the attribute, and doesn't implement the kItemFileBaseNameRole attribute to get it.

Log: Optical drive list view, after mounting, it takes a while for the file names inside the CD-ROM to be displayed in their entirety
Bug: https://pms.uniontech.com/bug-view-225935.html