fix: CD-ROM drive opens for the first time with a sorting error
Create fileinfo for optical drive on first open, asynchronous fileinfo agent created in optical drive. The asynchronous optical drive agent fileinfo has not finished querying at the time of sorting, so the attribute is wrong. Modify to use dfmio's iterator to create an asynchronous fileinfo and get all the fileinfo properties when the optical drive iterator fetches the fileinfo.

Log: CD-ROM drive opens for the first time with a sorting error