fix: [device] lvm group not hidden.
the disks are fully-encrypted, after unlocked on other host, the lvm
group still display in computer page, and it's not mountable, which
should be hidden. the INSTALLER make a udev rule to tag it as HintIgnore
so that computer page would not show that in it's original host, but
when the device is plugined in other host, the rule file is not exists
and the INSTALLER has no chance to update the rule file, dfm should hide
it by default.

Log: fix issue about lvm device display.

Bug: https://pms.uniontech.com/bug-view-231653.html