fix: Drag a large file to the system disk, but there is no prompt indicating insufficient disk space. The file manager crashes directly
Check the disk space usage using the Gio interface. The space provided by Gio is sufficient, but in reality, it is not enough. Copying large files also uses mmap, so there will be no errors and crashes directly. Modify the unified device to obtain free space on the device

Log: Drag a large file to the system disk, but there is no prompt indicating insufficient disk space. The file manager crashes directly
Bug: https://pms.uniontech.com/bug-view-233019.html