fix: Copying a large number of empty files and folders on the USB drive, with incorrect progress bar statistics
The reason for the inaccurate data statistics here is that we calculated the size of the directory as 4096, but in the write statistics of the USB drive, we calculated the size by reading and writing how many block blocks were written. The size of the directory and the 0kb file were calculated as a numerical value in the USB drive, which caused our own calculated values to not match. When modifying the statistics size, we are copying thread blocking statistics and then using our own statistics method.

Log: Copying a large number of empty files and folders on the USB drive, with incorrect progress bar statistics
Bug: https://pms.uniontech.com/bug-view-236707.html