fix: [io/proxyinfo]Copy the application icon to the safe and paste it without displaying it. You need to manually refresh it
After updating the proxy's fileinfo, the proxy's fileinfo was not notified,When creating a monitor, when entering the directory, rootinfo creates a URL with a "/" end, while in infocache, a URL without a "/" end is used. There are two URLs in the same directory, so one is stopped and not all are stopped. Therefore, the safe is locked and unlocked, and the monitor becomes invalid.

Log: Copy the application icon to the safe and paste it without displaying it. You need to manually refresh it
Bug: https://pms.uniontech.com/bug-view-240677.html