fix: [url history] cannot remove history if url is punycode-ed
if url's host contains Non-ASCII char, it will be encoded by punycode,
which is like xn--xxxx, and the url about to mount is this kind of, when
mount failed, cannot remove the smb://xn--xxxx because it's not existed
in history list, before remove the history item, the xn--xxx should be
decoded with punycode.

Log: as title.

Bug: https://pms.uniontech.com/bug-view-237183.html