fix: In the blank space of the dde-file-manager or dde-desktop, using the shortcut key "Paste" in the right-click menu does not take effect
After using the shortcut key and executing the trigger function, after an interval of 50ms, the window active and focus events were received, resulting in invalid reading of the cached copy data and exiting the editing state. Delay processing of these two shortcut key events.
Log: In the blank space of the dde-file-manager or dde-desktop, using the shortcut key "Paste" in the right-click menu does not take effect
Bug: https://pms.uniontech.com/bug-view-244903.html