fix: The shortcut file "bensound sunny. mp3" in the music directory cannot be moved to the trash
Determine if there is an interface issue that supports the trash operation, and supplement the judgment. If it is a linked file and it is in the main directory, it can be deleted

Log: The shortcut file "bensound sunny. mp3" in the music directory cannot be moved to the recycle bin
Bug: https://pms.uniontech.com/bug-view-259171.html