fix: In the file manager, name the file with "." and the file manager will crash
Calling this slot function _q_commitDataAndCloseEditor will submit the edited content and perform the destruction of the editorwidget. If executed simultaneously, the first execution is completed before executing the deconstruction, and the second execution will crash. When it is executing, it is no longer processed.

Log: In the file manager, name the file with "." and the file manager will crash
Bug: https://pms.uniontech.com/bug-view-263687.html