fix: [desktop] double free of typeclassifier.
TypeClassifier instance was deleted in NormalizeMode, GeneralModelFilter
was deleted in TypeClassifier, when TypeClassifier instance passed into
GeneralModelFilter as a smart pointer, the managed instance will be also
deleted when the ref-count reduced to zero. but TypeClassifier is not a
`QEnabledFromThis` instance. and `Double free` occured.

the GeneralModelFilter only need manager nude pointer, though it's more
safe to manage a smart pointer.

Log: as title.

Bug: https://pms.uniontech.com/bug-view-264655.html