fix: Copy and paste the file to the root directory of FTP mount with an error message. After switching directories, an unknown file will be generated upon return
Received FTP creation information, but FTP remote creation failed as the file does not exist. Modify to check if a file exists when adding a child

Log: Copy and paste the file to the root directory of FTP mount with an error message. After switching directories, an unknown file will be generated upon return
Bug: https://pms.uniontech.com/bug-view-269893.html