fix: no permission to execute dde-file-manager on NixOS
log: In the new version of cmake, install FILES will automatically delete the executable permission, so the PERMISSIONS field must be declared.
A better way is use PROGRAMS form, it is identical to the FILES form except that the default permissions for the installed file also include OWNER_EXECUTE, GROUP_EXECUTE, and WORLD_EXECUTE. This form is intended to install programs that are not targets, such as shell scripts. Use the TARGETS form to install targets built within the project.

https://cmake.org/cmake/help/latest/command/install.html