fix: When taking a screenshot, select "Specify Save Location" and choose another directory (not the current directory). Changing the file name has no effect and the file name is highlighted. However, you still need to click on the file name again with the mouse to enter it
Focus setting issue, set the edit box to focus when highlighted

Log: When taking a screenshot, select "Specify Save Location" and choose another directory (not the current directory). Changing the file name has no effect and the file name is highlighted. However, you still need to click on the file name again with the mouse to enter it
Bug: https://pms.uniontech.com/bug-view-273811.html