fix: Fixed the crash issue of dragging files in desktop organization.
In Qt6, when the size of the list is exceeded, the insert operation of QList will trigger an assertion crash
To avoid the crash, if the size bigger than list's size, use append instead.

log: If the size bigger than list's size, use append instead.
bug: https://pms.uniontech.com/bug-view-284723.html
Influence: Dragging files in desktop organization.