fix: Fixed the issue that the search input box crashed
QCompleter responds to the input and triggers the ontextchanged operation, resulting in a new operation that has not been executed last time, so that the data in the list to be processed in the previous time has been cleared, and the second response operation is processed, and the first one is continued, at this time, the data is abnormal, and the data managed in the list has been released, resulting in no data in the data area pointed by QString at this time, and a null pointer is called.

Use Qt::QueuedConnection to connect the signal slots and let them process in order.

log: AUse queueconntion to connect the signal slots and let them process in order.
bug: https://pms.uniontech.com/bug-view-283947.html
Influence: The normal use of the search input box.