fix: Fixed the issue that deleting a valut and creating a new one failed.
When using QProcess pointer variables, changing the environment variables will have an effect on the entire life cycle,
and thus will also be affected when other processes are started. Each change to the environment variables appends the system environment variables instead of replacing them.

In this bug, starting cryfs when mounted on a directory will change the environment variables, resulting in the loss of some system environment variables when using cryfs-unmount.

log: The lack of environment variables causes cryfs-unmount to be unable to find the corresponding path, resulting in the failure of folder unmount.

bug: https://pms.uniontech.com/bug-view-284441.html
Influence: The normal use of the vault.