fix: Fixed the issue where desktop files with special characters and mixed Chinese and English characters in the file name could not be opened.
The dde-file-manager script file parses mixed Chinese and English and special characters in the form of ascall, resulting in incorrect encoding when encoding multi-byte, single-byte and special characters into URLs.
Here it is modified to parse in the form of utf-8 character encoding.

log: Modify the encoding method from chat-gpt.
bug: https://pms.uniontech.com/bug-view-283841.html
Influence: Focus on the bug itself and whether it introduces other unknown problems.