fix: Fixed the transparent encryption of the vault. Unlocking failed when the safe was opened again after the screen was locked.
When using g_unix_mount_get_fs_type to obtain a symbolic link, it is not resolved into a real path, resulting in an error in obtaining the file mount type.
Therefore, convert the symbolic link into a real file path before using it.

log: Get the mount type of the vault
bug: https://pms.uniontech.com/bug-view-278715.html