fix: Fixed the issue where desktop application shortcuts could not be deleted by dragging them to the Recycle Bin window of the file manager
During the dragging process, the desktop application shortcut cannot be deleted due to its non-renameable attribute. When determining whether it can be dragged, determine whether it is dragged to the Recycle Bin. If it is dragged to the Recycle Bin, it can be deleted by dragging.

log: Fixed the issue where desktop application shortcuts could not be deleted by dragging them to the Recycle Bin window of the file manager

bug: https://pms.uniontech.com/bug-view-285659.html