fix: Fix the problem of deformation of the crumb menu.
The rounded corner set by the system is too large. When it is displayed only one item, it cannot be completely drawn to the corner, causing the lower part of the window to abnormal. Compatible treatment will increase height when there is only one item.

Log: Fix the problem of deformation of the crumb menu.

Bug: https://pms.uniontech.com/bug-view-296751.html