fix: Fixed the issue of not entering the username and password during mount, and directly closing the authentication window, prompting for incorrect information.
Use the cancel operation of the user manually closing the window and clicking the dialog box as a cancel operation.
Close the dialog box directly after cancel.

log: Fixed the issue of not entering the username and password during mount, and directly closing the authentication window, prompting for incorrect information.

Bug: https://pms.uniontech.com/bug-view-301791.html