fix: add multilingual support for empty trash action
Expanded the localization of the "Clean Trash" action in the dde-trash.desktop file by adding translations for multiple languages, including but not limited to Arabic, Chinese, Spanish, and Russian. This enhancement improves accessibility and usability for a broader audience.

Log: Added translations for "Clean Trash" in various languages to enhance user experience.

bug: https://pms.uniontech.com/bug-view-329941.html